+++ common/src/main/java/net/conczin/immersive_furniture/item/FurnitureItem.java
@@ -2,11 +2,13 @@ package net.conczin.immersive_furniture.item;
 
 import net.conczin.immersive_furniture.block.*;
 import net.conczin.immersive_furniture.data.FurnitureData;
+import net.conczin.immersive_furniture.data.FurnitureDataManager;
 import net.conczin.immersive_furniture.data.ServerFurnitureRegistry;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.BlockItem;
@@ -19,10 +21,8 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.FluidState;
 import net.minecraft.world.level.material.Fluids;
 
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
+import java.util.concurrent.ConcurrentSkipListSet;
 
 public class FurnitureItem extends BlockItem {
     public static final String FURNITURE = "Furniture";
@@ -103,6 +103,8 @@ public class FurnitureItem extends BlockItem {
         return true;
     }
 
+    private static final Set<String> alreadySaved = new ConcurrentSkipListSet<>();
+
     @Override
     protected boolean placeBlock(BlockPlaceContext context, BlockState state) {
         // First place the main block
@@ -140,6 +142,13 @@ public class FurnitureItem extends BlockItem {
         // Keep track of placed furniture to estimate usage
         if (level instanceof ServerLevel serverLevel) {
             ServerFurnitureRegistry.increase(serverLevel, getData(context.getItemInHand()));
+        }
+
+        // Also explicitly save the data to fix possible save issues
+        String hash = data.getHash();
+        if (!alreadySaved.contains(hash)) {
+            alreadySaved.add(hash);
+            FurnitureDataManager.save(data, new ResourceLocation("hash", hash));
         }
 
         return true;