+++ common/src/main/java/net/conczin/immersive_furniture/network/s2c/FurnitureDataResponse.java
@@ -21,6 +21,6 @@ public record FurnitureDataResponse(String hash, FurnitureData data) implements 
 
     @Override
     public void handle(Player e) {
-        FurnitureDataManager.save(data, new ResourceLocation("hash", hash));
+        FurnitureDataManager.save(data, new ResourceLocation("cache", hash));
     }
 }